requirements.txt
aiogram==3.4.1
asyncpg
exifread
aiohttp
apscheduler
openai
python-dotenv